pyproject.toml
[project]
name = "homelab-manager"
version = "0.1.0"
description = "Homelab Manager project to manage different things on my homelab server(s)"
authors = [
    {name = "Abdulrahman Mazhar",email = "whoamimaohw.career@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yt-dlp (>=2025.7.21,<2026.0.0)",
    "ffmpeg (>=1.4,<2.0)",
    "pony (>=0.7.19,<0.8.0)",
    "aio-pika (>=9.5.5) ; python_version >= '3.13' and python_version < '4.0'",
    "testcontainer-python-rabbitmq (>=0.4.0,<0.5.0) ; python_version >= '3.13' and python_version < '4.0'",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
packages = [
    {include = "src", from = "."}
]


[tool.poetry.group.dev.dependencies]
ruff = "^0.12.7"
black = "^25.1.0"
ipykernel = "^6.30.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
coverage = "^7.10.2"

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]